Preserve REST access token across application restarts
Make RestAdapter store the access token in NSUserDefaults, which is
a per-app shared persistent storage and typically used for keeping app
preferences and such.

Add a test to check if the access token is stored in the app-wide shared
persistent storage -- if the token is successfully stored and reused,
the value will be shared among independently constructed adapters.