Fix implementation issues found in LBFile and SLRESTAdapter. Add a feature to the adapter's invoke methods for handling response data via output stream.
LBFile and SLRESTAdapter seemed left in a temporary implementation to get
the file upload/download working. (e.g. the contract for file download needed
to be "/%@/:container/download" with multipart specified, which should have
been "/%@/:container/download/:name" w/o multipart).

In order to support file downloading in an efficient manner, adapter's invoke
methods now include the versions that takes an output stream as a parameter.
Those are used in file download in LBFile.  Those are also intended to be used
for binary payload download to a data buffer in future.

Multipart form handling was made better (it used to be specialized just for
the purpose of file uploading).  If multipart is set to YES in the contract,
SLRESTAdapter's invoke methods construct each part based on a StreamParam
object supplied via the parameters.

Made SLRESTContractItem's instance initialization methods private, as those
won't be called directly by developers.